Update build artifact workflows & repository messaging (#4797)
## Explanation
This PR does two things:
- It updates repository messaging to indicate multiple team members may
be unavailable until early January (which will show on all new/reopened
PRs once this is merged). I'll update existing PRs so that authors are
aware.
- It updates the build tests CI workflow to only upload the Oppia dev
APK (as this is the only one practically useful for developers since the
AABs can't be directly installed on a device).

## Essential Checklist
- [ ] ~The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)~
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- Infrastructure-only change.