Fix in-lesson language switching.
This also adds Swahili translations for test topic 0. Some style changes
including support for a secondary button type. Updated the in-lesson
language switcher to use this style & to be gated behind the learner
study platform parameter.

And, fixes hint & solution header expanding to be the whole header and
not the whole card.