Fix #4765: The Continue animation isn't interactive (#4814)
<!-- READ ME FIRST: Please fill in the explanation section below and
check off every point from the Essential Checklist! -->
## Explanation

Fixes https://github.com/oppia/oppia-android/issues/4765

Demo Video:


https://user-images.githubusercontent.com/64526117/210055420-016f2dd2-46d6-4166-ae2a-be4c98710299.mp4

This PR tries to handle the problem of the continue interaction button's
animation not being eye-catching to the users. Currently, it does so by
changing the animation to a shake animation. The button now does a
little wobble, starting 45s after the screen is shown. It wobbles 2
times, and then repeats the animation after every 8 seconds.

Only with further testing and trial, will we know how the animation
appeals to the users and resolves the problem at hand.

**Technical detail:**
The anim file (xml) is set to play the animation 2 times. 
In the `ContinueButtonView.kt` file, the startAnimating() function has
been made recursive, which calls itself after every 8 seconds in order
to repeat the animation. The 8 second gap is achieved using the
lifeCycleSafeTimerFactory.

## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only