First commit of new asset download script.
This is utilizing a new Oppia web API being worked on that will allow
for an easier collection of Oppia assets as part of a broader initiative
to fully automate Oppia Android's release pipeline.

There are a bunch of TODOs, documentation, testing, and functionality
work needed yet before this solution is ready to check in (including
introducing new proto to legacy proto conversion & outputting for the
app to actually be able to consume the lessons being imported from Oppia
web).

This also sets the groundwork for downloads infrastructure both for the
client & Oppia web's new proto APIs that will be introduced as part of
itnroducing lessond download support in the app.