Migrates to Bazel 5.0 & Kotlin 1.6.
This is early work to migrate the codebase to Bazel 5.x (and maybe 6.x),
as well as Kotlin to 1.6 (and maybe 1.7). The Kotlin migration is needed
due to an incompatibility with older rules_kotlin and Bazel 5.x. Other
dependencies needed updating, as well, including Dagger and rules_java.
Issue #1535 is being fixed as part of these upgrades.

As part of working through the Dagger issues that were encountered, it
was noticed that the codebase was actually setting up Dagger
incorrectly. This has been corrected: the codebase now defines a single
top-level dagger_rules().

This commit is not yet the final solution--the build doesn't yet work
and requires investigation before the rest of the work can be completed.

This commit partly continues the migration work from #4092 though it's
not yet finished. What remains includes:
- Migrating to latest Bazel & Kotlin versions
- If possible, modularize all third-party deps like oppia-proto-api
- Enable strict Kotlin & Java deps
- Enable warnings-as-errors for Kotlin & Java (and fix all issues)
- Ensure Guava-Android is used instead of JRE
- Resolve all Maven conflicts
- Remove dependence on Oppia's custom Android tools
- Remove support on legacy D8 (needs verification)
- Remove disabling incremental dexing (for faster builds)
- Disabling header compilation being disabled (for faster builds)