Main follow-up alpha MR6 fixes.
This does a lot of stuff (that will be more clearly represented in the
corresponding PR), but it includes:
- An updated learner profile screen that includes a clearer sync status,
a share button that also shares event logs stats and the logs
themselves, event stats, and a button to force log uploads.
- Fix to PersistentCacheStore that prevents potential log duplication.
- Fix to AnalyticsController that could result in logs being lost when
they're supposed to be uploaded.
- Long-term persistence of events that have been uploaded to ensure that
they can be tracked & shared.