Fix #4850: Bump version codes for alpha 0.10 release, and fix version code ordering (#4851)
## Explanation
Fixes #4850

This PR introduces the necessary version code updates for an upcoming
0.10 alpha release (2 different RCs). It also fixes the order of version
codes such that less stable versions of the app (i.e. build flavors)
take priority over more stable ones (so that users signed up for an
earlier pre-release will get that over later versions, e.g. users would
receive an alpha version of the app over beta if they're in both). This
happens because Play Store uses version codes as the actual comparable
versions for uploaded binaries, and it picks the binary with the highest
version code to be installed based on all that a user may qualify for
(based on which release tracks that they're participating in). The
assumption here is that newer versions maintain backward compatibility
with older versions, and this is something we will continue to aim for
(though with minimal guarantees for pre-release versions).

Note that one unfortunate side effect of this change is that local
installs of the app will require manually uninstalling when installing a
"more stable" build flavor (though technically this was an issue in the
other direction previously). This does make "upgrade" flows a little
trickier to test locally unless using pre-assembled binaries.

This PR also updates app versioning such that the version now includes
the local branch's commit rather than the common commit with develop (to
ensure RC rebuilds have properly unique version names rather than
sharing one version name for the release branch).

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This is an infrastructure-only change.