Gate long-term event tracking.
This commit gates tracking events as 'uploaded' behind the learner study
analytics feature (since that feature exposes the learner analytics
page, it's the only time that tracking events in this way is actually
beneficial).

Tests have been updated & added to ensure that the functionality is
correctly gated behind the feature's platform parameter.