Fix actual compute_affected_tests issue.
The problem was that long outputs (such as when WORKSPACE is included in
the changes) can cause Bazel to hang if its standard output isn't being
read. This commit introduces a new implementation for CommandExecutor
that ensures both a timeout and the process's output can be observed
simultaneously. This actually should result in slightly better
performance across all scripts.

Additionally, a few correctness and optimization changes were made to
the queries around finding tests affected by build file changes. This
has resulted in about a 4x improvement in speed for the script. There's
one more potential optimization that could reduce this by almost 10x
more, but it seems to result in incorrect results so it would require
further investigating.

Due to the time & correctness issues, the timeout for
ComputeAffectedTests-related processes has been reduced from 30 minutes
to 5 minutes. It's still kept over the default 30 seconds since some of
these commands can be quite time consuming on CI (and get increasingly
more expensive as the build graph increases in size over time). This
adjustment should keep the script future-proof for a while.