Fix broken tests.
For some reason, the old kotlinx-coroutines-test was causing a
codebase-wide incompatibility with the 1.6.x kotlinx-coroutines-core
library, so this updates that version.

It also updates Gradle dependencies for consistency, and to fix the
Gradle builds.

And, this fixes the IntelliJ Bazel plugin syncing issue when using
kt_jvm_import (for kotlinx-coroutines-core-jvm).