Fix part of #2116: revert "#4874 Move PromotedStoryListAdapter to BindableAdapter" (#4951)
This reverts commit `7fd290d68f0440d926f2d443dbd7bfb28ab20547`, fixing
part of #2116.

While testing #2116, we found that the UI for default profile view does
not display correctly, and traced the change to #4874 in [this
discusson](https://github.com/oppia/oppia-android/issues/2116#issuecomment-1512624791).

#4874 introduces a change to `MarginBindingAdapters` which utilizes
`MarginLayoutParamsCompat` to compute the margins of a layout before
drawing it. I suspect that this works for the `Promoted Story` and
`Topic Summary` because we compute the start and end margins of each
item relative to grid columns laid out on the HomeActivity, but no such
implementation exists for the Profile Chooser view. We need to do some
further investigation into this, hence the decision to revert.

We will need to test all the associated screens to make sure they still
look as expected:

![Screenshot 2023-04-19 at 18 01
28](https://user-images.githubusercontent.com/59600948/233117952-8c981fa2-2d0c-45cc-80ff-651862c1fd96.png)


## Essential Checklist
<!-- Please tick the relevant boxes by putting an "x" in them. -->
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
This was mainly a re-implimentation per #632, and the UI did not change.
Please refer to #4874 for screenshots.

The Affected screen displays okay with this revert:


![Screenshot_1681916915](https://user-images.githubusercontent.com/59600948/233125096-2242fb7e-eb9f-4b78-8db0-c22abb6b5cc3.png)

Other screens:
|   |   |
|---|---|

|![Screenshot_1681917117](https://user-images.githubusercontent.com/59600948/233126283-45a860c7-26ac-47f6-8b38-e54b1ee7cb1f.png)|![Screenshot_1681917190](https://user-images.githubusercontent.com/59600948/233126350-08033853-4434-462d-9bf3-98040ff2356d.png)|