Customize rules_kotlin.
This customizes rules_kotlin to support a variety of fixes to make both
strict and unused deps viable for the Oppia Android codebase (though
much more of the changes benefit unused deps checking).

This includes significant patched-in customizations to rules_kotlin and
changes the build pipeline such that rules_kotlin itself is built prior
to using it. This is a small add-on in build time (particularly
noticeable for initial builds) that the team will have to cope with
until the patched changes can be accepted into rules_kotlin itself &
released.

This commit also enables actual error-based checking for both strict and
unused deps.