De-couple high-level app structures.
This essentially decomposes "app" & related libraries into much more
targeted libraries for activities, fragments, views, listeners, and view
models. It's a pivotal step in performing the broad production
modularization that will be occurring in a later PR.

There's a major runtime component to this change: it requires decoupling
injectors for views, fragments, and activities from their corresponding
Dagger component implementations. This converts a previous compile-time
type check to now a runtime check which could fail in the future. Moving
to Dagger Hilt will fully solve this as it will generate the necessary
injectors and components at compile-time to ensure proper compile-time
type resolution.