Add structured target parsing and improve analysis.
This improves upon target handling & parsing in SuggestBuildFixes by
introducing proper structured parsing. This mainly improves confidence
in the behavior of the utility, but it also allowed for a trivial
introduction of sorting that matches buildifier's order (so no more
needing to run buildifier after auto-fixing a dependency list).

This new structure may have future benefits, as well, if it were to be
integrated into BazelClient.

Besides the parsing, this also updates SuggestBuildFixes to only analyze
targets which explicitly have issues after attempting to build provided
patterns. This *significantly* improves script performance (to the point
where it may be viable to incorporate into an incremental workflow).

It *may* be possible to also now incorporate the script into a CI check
but building all targets may still be prohibitively expensive.