Introduce new Bazel setup wizard.
This is an all-in-one initialization experience for users to ensure
their local environment is both working for builds, and that they're set
up to be able to start building/changing the codebase. While the actual
setup process can be lengthy (~hours), the wizard is designed to
carefully monitor progress from constituent steps to ensure that
everything is correctly working. Also, it's only expected that people
need to run this flow once per new repository clone. Advanced users can
skip the sanity checking and just generate (or regenerate) the necessary
files.

The only enforcement for running the setup flow currently is the
generation of the .bazelproject file needed to import the project, so
some thought will also be put into whether this check can be made more
strict to also cover users who aren't using Android Studio (or try to
import the project directly).