Fix issue with compute stats workflow.
If the Bazel version changes then the previous binaries won't build, so
this changes the stats workflow to install both the previous and current
versions of Bazel. While this is only needed for the PR making the
version change, it should be low-impact to always include (plus, Bazel
updates often so these changes might become more frequent in the
long-term).