Bunch of KDoc fixes.
This addresses a few remaining areas after a follow-up review of #4762.
It also systematically ensures a variety of KDocs issues can't be
introduced in the future by adding 8 new regex checks + tests (with a
bunch of style fixes throughout the codebase). While a lot of this is
unrelated to the PR at hand, it's nice to just get this done in a PR
that's about to get merged.

Some other style things were done in this commit, as well.