Fix #5005: Reorganize CODEOWNERS for clearer segmentation & to use GitHub groups (#5006)
## Explanation
Fixes #5005.
Fixes part of #5004.

This PR completed reorganizes codeowners in two ways:
1. It tries to more aggressively segment directories based on common
themes. Longer term (per #5004) the goal is to eventually organize the
codebase itself based on features and infrastructural components, but
that requires much more that needs to be done yet.
2. It moves all codeownership over to four teams: @oppia/owners,
@oppia/android-app-infrastructure-reviewers,
@oppia/android-dev-workflow-reviewers, and
@oppia/android-frontend-reviewers

This PR does attempt to keep the divisions more logical than necessarily
even. At a high-level:
- oppia/owners own anything that's especially sensitive and requires
additional care or should never really be changed.
- oppia/android-frontend-reviewers own everything tied to code that
directly affects the end-user experience.
- oppia/android-app-infrastructure-reviewers own everything that both
supports and defines domain logic.
- oppia/android-dev-workflow-reviewers own everything that supports
making development easier, or is tied to central developer workflows
(such as the build infrastructure).

The above don't exactly match to existing teams. However, it's generally
assumed that:
- The team TL will generally review anything that goes to oppia/owners.
- The CLaM team is ultimately responsible for everything that falls
under oppia/android-frontend-reviewers and
oppia/android-app-infrastructure-reviewers. The distinction between
these two types of parts of the codebase allow for more granular
specialization as the knowledge required in each of these coding domains
tends to be different.
- The developer workflow team is ultimately responsible for everything
that falls under oppia/android-dev-workflow-reviewers-reviewers.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- this PR is only changing codeowners (which doesn't affect the
actual app or any builds/configurations).