Follow up CI fixes and cleanup.
This removes some no-longer-needed comments from AndroidLocaleFactory.
It also updates the factory to be a bit more robust against missing
region definitions (rather than trying to proceed which could result in
an exception being thrown when it doesn't need to be).

This also fixes InitializeDefaultLocaleRule incorrectly creating an
empty region in cases when no region data is provided. This change fixes
the StateFragmentLocalTest failures reported by CI.