Fix pipeline issues & add image conversion.
This fixes #5026 and #5023.

The first issue (story thumbnails not loading plus some other images not
working as expected) is caused by several different incompatibilities
that can occur in current lesson SVGs. Specifically, embedded bitmap
images are not supported by AndroidSVG. These script changes add
automatic incompatibility detection (though this can be improved in the
future) and render out a fixed bitmap to replace all references to the
broken SVG with the new bitmap. The resolution of this bitmap is
calculated based on the self-reported size of the image (via its
filename) interpreted as interpreted "Oppia DiPs" based on the most
common device PPIs used by the app users. This produces a slightly fuzzy
image, but definitely passable until SVGs can be used permanently.

The second issue is addressed by fixing the output formatting for
multiple choice, item selection, and drag & drop choices to correctly
nest the list of subtitles into custom schema objects (in the same way
the JSON lessons do since that's what the current UI components expect).
This could be cleaned up, but it seems simpler just to match the
existing behavior to avoid changing the app more than necessary (plus,
this will automatically get addressed when the app is moved over to the
new lesson proto format).